Removed plots and metrics table.
Realised why the descriptor loss may be mostly flat...
Not passing any coordinate information back to the descriptors in the model, I just concatenate the tensors
rather than multiplying or adding. Assume that descriptors certainly need coordinate information as their loss
is based on location. Also assume that coordinates would benefit from descriptor value information as (in theory
and at a basic level), components lower down are more likely to be wheels, where as higher up are more likely to
be sensors (though this is an example and assumption about the robots, due to them being evolved, this may not
be the case, though I would assume some descriptor information would be beneficial to components).

Added another 1Dconv layer to the encoder for the descriptor mlp to make it the same size as the coordinate mlp output.
Replace the concatentation with additive combination and a LeakyReLU activation. Changed the Linear steps in the fc (refining sequential) to be smaller.
Added another intermediary layer in the decoder fc (upsampling) before split to mirror.

Scheduler set to none for quick runs.
Setup for some quick exploratory runs.